connectivity issues
for some reason my ui naviagtion won't change its appearance or my UIImage change wehn prompted. I don't understand.
