+++ Sprint 4 Challenge/Sprint 4 Challenge/ViewController.swift
@@ -2,27 +2,31 @@
 //  ViewController.swift
 //  Lock(Sprint 4)
 //
-//  Created by Jordan Davis on 5/30/19.
+//  Created by Jordan Davis on 5/31/19.
 //  Copyright © 2019 Jordan Davis. All rights reserved.
 //
 
 import UIKit
 
-class ViewController: UIViewController {
+import UIKit
 
+class ViewController: UIViewController {
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         imageView.image = UIImage(named: "Locked")
         lockButton.title = ""
         setAppearance()
+        AppearanceHelper.setCustomAppearance()
     }
+    
     
     @IBAction func ballValueChanged(_ sender: CustomControl) {
         if customControl.isUnlocked {
             imageView.image = UIImage(named: "Unlocked")
             customControl.isUserInteractionEnabled = true
             lockButton.title = "Lock"
-            navigationItem.title = "Reset"
+            navigationItem.title = "Unlocked!"
         }
     }
     
@@ -30,7 +34,7 @@ class ViewController: UIViewController {
         customControl.reset()
         imageView.image = UIImage(named: "Locked")
         lockButton.title = ""
-        navigationItem.title = "Slide to Unlock"
+        navigationItem.title = "Slide To Unlock"
     }
     
     func setAppearance() {
@@ -42,6 +46,7 @@ class ViewController: UIViewController {
         overallView.layer.shadowOpacity = 1
     }
     
+    
     //MARK: - Properties
     
     @IBOutlet weak var overallView: UIView!
@@ -51,5 +56,15 @@ class ViewController: UIViewController {
     @IBOutlet weak var lockButton: UIBarButtonItem!
     
 
+}
+
+extension UIColor {
+    
+    convenience init(_ red: CGFloat, _ green: CGFloat, _ blue: CGFloat, _ alpha: CGFloat = 1) {
+        self.init(red: red/255, green: green/255, blue: blue/255, alpha: alpha)
+    }
+    
+    static let smokeGray = UIColor(214, 214, 214)
+    static let marvelRed = UIColor(137, 10, 24)
 }
